requirements.txt
aiohttp==3.4.4
aiohttp-jinja2==1.1.0
aiomysql==0.0.19
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
idna==2.7
idna-ssl==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multidict==4.4.2
mysqlclient==1.3.13
pycparser==2.19
PyMySQL==0.9.2
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.14
yarl==1.2.6
